# gestion-soutenance-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion des Soutenances",
        "description": "Documentation OpenAPI pour le backend Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/admin/soutenances": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Lister les soutenances",
                "operationId": "633dad0d19aa2bae61af959c5f87364b",
                "responses": {
                    "200": {
                        "description": "Liste des soutenances"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Planifier une soutenance",
                "operationId": "f1c43100c52b779df850cb757bb56d18",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "heure",
                                    "salle",
                                    "etudiant_id"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Soutenance créée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/soutenances/{soutenance}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Mettre à jour une soutenance",
                "operationId": "a0694f3f508d04e1fdc8cb136de39928",
                "parameters": [
                    {
                        "name": "soutenance",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/soutenances/{soutenance}/note": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Mettre à jour la note finale",
                "operationId": "95d4e3e32e14bd6e61c1c59f91b90529",
                "parameters": [
                    {
                        "name": "soutenance",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "note_finale"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Saisir uniquement le jeton d'accès",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}